--- a/UefiCpuPkg/Library/MpInitLib/MpEqu.inc
+++ b/UefiCpuPkg/Library/MpInitLib/MpEqu.inc
@@ -74,18 +74,18 @@ struc MP_CPU_EXCHANGE_INFO
   .StackStart:                   CTYPE_UINTN 1
   .StackSize:                    CTYPE_UINTN 1
   .CFunction:                    CTYPE_UINTN 1
+  .NumApsExecuting:              CTYPE_UINTN 1
+  .ApIndex:                      CTYPE_UINTN 1
   .GdtrProfile:                  CTYPE_UINT8 IA32_DESCRIPTOR_size
   .IdtrProfile:                  CTYPE_UINT8 IA32_DESCRIPTOR_size
   .BufferStart:                  CTYPE_UINTN 1
   .ModeOffset:                   CTYPE_UINTN 1
-  .ApIndex:                      CTYPE_UINTN 1
   .CodeSegment:                  CTYPE_UINTN 1
   .DataSegment:                  CTYPE_UINTN 1
   .EnableExecuteDisable:         CTYPE_UINTN 1
   .Cr3:                          CTYPE_UINTN 1
   .InitFlag:                     CTYPE_UINTN 1
   .CpuInfo:                      CTYPE_UINTN 1
-  .NumApsExecuting:              CTYPE_UINTN 1
   .CpuMpData:                    CTYPE_UINTN 1
   .InitializeFloatingPointUnits: CTYPE_UINTN 1
   .ModeTransitionMemory:         CTYPE_UINT32 1
@@ -100,5 +100,14 @@ struc MP_CPU_EXCHANGE_INFO
   .SevSnpKnownInitApicId:        CTYPE_BOOLEAN 1
 endstruc
 
-MP_CPU_EXCHANGE_INFO_OFFSET equ (Flat32Start - RendezvousFunnelProcStart)
+;
+; Declare a UINTN struct for the purpose of
+; of obtaining the size of a UINTN (UINTN_size).
+;
+struc UINTN
+  .Data                          CTYPE_UINTN  1
+endstruc
+
+; MP_CPU_EXCHANGE_INFO Offset (UINTN aligned)
+MP_CPU_EXCHANGE_INFO_OFFSET equ (((Flat32Start - RendezvousFunnelProcStart) + (UINTN_size - 1)) & ~(UINTN_size - 1))
 %define MP_CPU_EXCHANGE_INFO_FIELD(Field) (MP_CPU_EXCHANGE_INFO_OFFSET + MP_CPU_EXCHANGE_INFO. %+ Field)