# Description

UefiCpuPkg/MpInitLib: Fix split-lock violation from MP_CPU_EXCHANGE_INFO

With latest x86_64 processors and Linux kernel which support split-lock detection, OVMF was implicated
in a split-lock violation.  It was determined that the NumApsExecuting field of the MP_CPU_EXCHANGE_INFO
struct in MpInitLib (which is used atomically by the AP Reset Vector assembly code) was the culprit resulting in
warning from the kernel, like so:

    [1035537.770691] x86/split lock detection: #AC: qemu-kvm/785650 took a split_lock trap at address: 0x7fe40050

which aligns with associated OVMF debug output:

    AP Vector: non-16-bit = 7FE40000/DF
    WakeupBufferStart = 9F000, WakeupBufferSize = 0
    AP Vector: 16-bit = 9F000/41, ExchangeInfo = 9F041/5C



Therefore, align the MP_CPU_EXCHANGE_INFO structure and move the NumApsExecuting field
to the beginning to ensure it resides within a single cache-line avoiding the split-lock. Do the same
for the ApIndex field as it is also used atomically and thus subject to a potential split-lock violation.

Results of the fix (with added OVMF debug to output the address of NumApsExecuting field):

Before fix:

    AP Vector: non-16-bit = 7FE40000/DF
    WakeupBufferStart = 9F000, WakeupBufferSize = 0
    AP Vector: 16-bit = 9F000/41, ExchangeInfo = 9F041/5C
    MpCpuExchangeInfo->NumApsExecuting Addr = 9F07D

A 4 byte field at address 9F07D crosses a cache-line boundary resulting in a lock-split violation. Also, notice the ExchangeInfo struct is unaligned at addr 9F041
    
With fix:

    AP Vector: non-16-bit = 7FE40000/E5
    WakeupBufferStart = 9F000, WakeupBufferSize = 0
    AP Vector: 16-bit = 9F000/41, ExchangeInfo = 9F048/5C
    MpCpuExchangeInfo->NumApsExecuting Addr = 9F054

Aligned address 9F054 does not cross a cache-line avoiding the lock-split violation. And, notice the ExchangeInfo struct is now aligned at addr 9F048.

NOTE: The OVMF split-lock violations result in a non-trivial performance degradation (due to bus locking). Here is some
profiling data with and without this fix:

    VCPUS:  AP Finish Time (microseconds):          # kernel split-lock detections:
    (max)   (without fix)   (with fix)              (without fix)                   (with fix)
    2       22557           30                      1                               0
    4       44738           46                      3                               0
    8       76974           60                      6                               0
    16      124207          170                     10                              0
    32      128673          241                     ~15 (callbacks suppressed)      0
    64      233840          424                     ~30 (callbacks suppressed)      0
    128     417998          1161                    ~58 (callbacks suppressed)      0
    254     820083          7395                    ~114 (callbacks suppressed)     0

Cc: Ray Ni <ray.ni@intel.com>
Cc: Jiaxin Wu <jiaxin.wu@intel.com>
Cc: Zhiguang Liu <zhiguang.liu@intel.com>
Cc: Dun Tan <dun.tan@intel.com>
Cc: Rahul Kumar <rahul1.kumar@intel.com>
Cc: Gerd Hoffmann <kraxel@redhat.com>
Cc: Star Zeng <star.zeng@intel.com>

- [ ] Breaking change?
  - **Breaking change** - Does this PR cause a break in build or boot behavior?
  - Examples: Does it add a new library class or move a module to a different repo.
- [ ] Impacts security?
  - **Security** - Does this PR have a direct security impact?
  - Examples: Crypto algorithm change or buffer overflow fix.
- [ ] Includes tests?
  - **Tests** - Does this PR include any explicit test code?
  - Examples: Unit tests or integration tests.

## How This Was Tested

OVMF Ia32X64 and X64 builds were tested on latest Intel X9 hardware (which supports split-lock detection). With the fix applied, the split-lock violations no longer occur.

## Integration Instructions

N/A