@ajyoung-oracle 2025-05-19
> The code change looks generally good to me. I added one minor comment.
> 
> But the commit message is too short when comparing the description put in the PR. Can you please include more details that you put in the PR description in the commit message? At least please highlight two things:
> 
>     1. the MP_CPU_EXCHANGE_INFO structure is not allocated in an aligned address.
> 
>     2. NumApsExecuting/ApIndex are type of UINTN. GdtrProfile/IdtrProfile are 6 (32bit) or 10(64bit) bytes which may cause NumApsExecuting/ApIndex across cacheline boundary (assuming cache line size is 64byte). That's the reason we need to put the two fields before GdtrProfile/IdtrProfile.

Thanks Ray for the review. 

I updated the code to align the MP_CPU_EXCHANGE_INFO struct on a UINTN boundary (instead of explicit 8 bytes).

I also made the commit  message more verbose as you suggested above.

Thanks again.